fix: correct AxSecurityDataEntityPermission element order + dedupe builder
Element order: the DataEntityPermissions block copied the Name-first
layout from AxSecurityEntryPointReference, but the Microsoft serializer
emits AxSecurityDataEntityPermission children as Grant, Name, Fields,
Methods (Grant first), with <Grant> CRUD elements in alphabetical order
(Correct, Create, Delete, Read, Update). Verified against shipped
ApplicationCommon privileges AgentFeedEntityMaintain/View.xml. The prior
output (Name-first, Read/Create/Update/Delete/Correct) was non-canonical
and produced a noisy re-sort on first save in Visual Studio.

Dedupe: generateAxSecurityPrivilegeXml was byte-identical in the two
mirrored XmlTemplateGenerator classes (createD365File.ts and
generateD365Xml.ts) and had to be patched twice by hand. Extracted into
a shared pure builder (securityPrivilegeXml.ts) that both delegate to, so
they can no longer drift.

Test locks in the canonical element order so future edits can't regress it.
