+++ src/tools/createD365File.ts
@@ -18,6 +18,7 @@ import { enforceGrounding } from '../utils/provenanceStore.js';
 import { gateOnFormPatternErrors } from './validateFormPattern.js';
 import { gateOnReferenceErrors } from './resolveReferences.js';
 import { normalizeD365Xml } from '../utils/d365XmlNormalizer.js';
+import { buildAxSecurityPrivilegeXml } from './securityPrivilegeXml.js';
 
 /**
  * Per-project-file mutex to serialise concurrent addToProject calls.
@@ -2589,52 +2590,12 @@ ${relationsXml}
   }
 
   /**
-   * Generate AxSecurityPrivilege XML.
-   * properties.targetObject  – ObjectName of the target menu item (optional)
-   * properties.objectType    – MenuItemDisplay | MenuItemAction | MenuItemOutput (default: MenuItemDisplay)
-   * properties.accessLevel   – 'view' | 'maintain' | 'read' (default: 'view' = Read only)
-   * properties.dataEntity    – Name of the data entity to grant permissions on (optional)
+   * Generate AxSecurityPrivilege XML. Delegates to the shared builder so this
+   * mirror and the one in generateD365Xml.ts cannot drift.
+   * @see buildAxSecurityPrivilegeXml for the property contract and element order.
    */
   static generateAxSecurityPrivilegeXml(name: string, properties?: Record<string, any>): string {
-    const label = properties?.label || '@TODO:LabelId';
-    const targetObject: string | undefined = properties?.targetObject;
-    const objType: string = properties?.objectType || 'MenuItemDisplay';
-    const al = (properties?.accessLevel || 'view').toLowerCase();
-
-    let entryPointsXml: string;
-    if (targetObject) {
-      const grantXml = al === 'maintain'
-        ? '\t\t\t\t<Read>Allow</Read>\n\t\t\t\t<Update>Allow</Update>\n\t\t\t\t<Create>Allow</Create>\n\t\t\t\t<Delete>Allow</Delete>'
-        : '\t\t\t\t<Read>Allow</Read>';
-      entryPointsXml = `\n\t\t<AxSecurityEntryPointReference>\n\t\t\t<Name>${targetObject}</Name>\n\t\t\t<Grant>\n${grantXml}\n\t\t\t</Grant>\n\t\t\t<ObjectName>${targetObject}</ObjectName>\n\t\t\t<ObjectType>${objType}</ObjectType>\n\t\t\t<Forms />\n\t\t</AxSecurityEntryPointReference>\n\t`;
-    } else {
-      entryPointsXml = '';
-    }
-
-    const dataEntity: string | undefined = properties?.dataEntity;
-    let dataEntityPermissionsXml: string;
-    if (dataEntity) {
-      const grantXml = al === 'maintain'
-        ? '\t\t\t\t<Read>Allow</Read>\n\t\t\t\t<Create>Allow</Create>\n\t\t\t\t<Update>Allow</Update>\n\t\t\t\t<Delete>Allow</Delete>\n\t\t\t\t<Correct>Allow</Correct>'
-        : '\t\t\t\t<Read>Allow</Read>';
-      dataEntityPermissionsXml = `\n\t\t<AxSecurityDataEntityPermission>\n\t\t\t<Name>${dataEntity}</Name>\n\t\t\t<Grant>\n${grantXml}\n\t\t\t</Grant>\n\t\t\t<Fields />\n\t\t\t<Methods />\n\t\t</AxSecurityDataEntityPermission>\n\t`;
-    } else {
-      dataEntityPermissionsXml = '';
-    }
-
-    const dataEntityPermissionsElement = dataEntityPermissionsXml
-      ? `<DataEntityPermissions>${dataEntityPermissionsXml}</DataEntityPermissions>`
-      : '<DataEntityPermissions />';
-
-    return `<?xml version="1.0" encoding="utf-8"?>
-<AxSecurityPrivilege xmlns:i="http://www.w3.org/2001/XMLSchema-instance">
-\t<Name>${name}</Name>
-\t<Label>${label}</Label>
-\t${dataEntityPermissionsElement}
-\t<DirectAccessPermissions />
-\t<EntryPoints>${entryPointsXml}</EntryPoints>
-\t<FormControlOverrides />
-</AxSecurityPrivilege>`;
+    return buildAxSecurityPrivilegeXml(name, properties);
   }
 
   /**
