+++ src/tools/generateD365Xml.ts
@@ -10,6 +10,7 @@ import { z } from 'zod';
 import { getConfigManager } from '../utils/configManager.js';
 import { ensureXppDocComment, ensureBlankLineBeforeClosingBrace } from '../utils/xppDocGen.js';
 import { decodeXmlEntitiesFromXppSource } from './modifyD365File.js';
+import { buildAxSecurityPrivilegeXml } from './securityPrivilegeXml.js';
 
 const GenerateD365XmlArgsSchema = z.object({
   objectType: z
@@ -1379,46 +1380,10 @@ ${relationsXml}
 </AxMenuExtension>`;
   }
 
+  // Delegates to the shared builder so this mirror and the one in
+  // createD365File.ts cannot drift. @see buildAxSecurityPrivilegeXml.
   static generateAxSecurityPrivilegeXml(name: string, properties?: Record<string, any>): string {
-    const label = properties?.label || '@TODO:LabelId';
-    const targetObject: string | undefined = properties?.targetObject;
-    const objType: string = properties?.objectType || 'MenuItemDisplay';
-    const al = (properties?.accessLevel || 'view').toLowerCase();
-
-    let entryPointsXml: string;
-    if (targetObject) {
-      const grantXml = al === 'maintain'
-        ? '\t\t\t\t<Read>Allow</Read>\n\t\t\t\t<Update>Allow</Update>\n\t\t\t\t<Create>Allow</Create>\n\t\t\t\t<Delete>Allow</Delete>'
-        : '\t\t\t\t<Read>Allow</Read>';
-      entryPointsXml = `\n\t\t<AxSecurityEntryPointReference>\n\t\t\t<Name>${targetObject}</Name>\n\t\t\t<Grant>\n${grantXml}\n\t\t\t</Grant>\n\t\t\t<ObjectName>${targetObject}</ObjectName>\n\t\t\t<ObjectType>${objType}</ObjectType>\n\t\t\t<Forms />\n\t\t</AxSecurityEntryPointReference>\n\t`;
-    } else {
-      entryPointsXml = '';
-    }
-
-    const dataEntity: string | undefined = properties?.dataEntity;
-    let dataEntityPermissionsXml: string;
-    if (dataEntity) {
-      const grantXml = al === 'maintain'
-        ? '\t\t\t\t<Read>Allow</Read>\n\t\t\t\t<Create>Allow</Create>\n\t\t\t\t<Update>Allow</Update>\n\t\t\t\t<Delete>Allow</Delete>\n\t\t\t\t<Correct>Allow</Correct>'
-        : '\t\t\t\t<Read>Allow</Read>';
-      dataEntityPermissionsXml = `\n\t\t<AxSecurityDataEntityPermission>\n\t\t\t<Name>${dataEntity}</Name>\n\t\t\t<Grant>\n${grantXml}\n\t\t\t</Grant>\n\t\t\t<Fields />\n\t\t\t<Methods />\n\t\t</AxSecurityDataEntityPermission>\n\t`;
-    } else {
-      dataEntityPermissionsXml = '';
-    }
-
-    const dataEntityPermissionsElement = dataEntityPermissionsXml
-      ? `<DataEntityPermissions>${dataEntityPermissionsXml}</DataEntityPermissions>`
-      : '<DataEntityPermissions />';
-
-    return `<?xml version="1.0" encoding="utf-8"?>
-<AxSecurityPrivilege xmlns:i="http://www.w3.org/2001/XMLSchema-instance">
-\t<Name>${name}</Name>
-\t<Label>${label}</Label>
-\t${dataEntityPermissionsElement}
-\t<DirectAccessPermissions />
-\t<EntryPoints>${entryPointsXml}</EntryPoints>
-\t<FormControlOverrides />
-</AxSecurityPrivilege>`;
+    return buildAxSecurityPrivilegeXml(name, properties);
   }
 
   static generateAxSecurityDutyXml(name: string, properties?: Record<string, any>): string {
